Fix to #9223 - Query: ArgumentNullException when using ternary operator with null comparison in the test expression
Problem was that try to optimize conditionals with logical iftrue/iffalse but the logic around but the logic was not robust enough. We can return null from Invert(...) method which throws the exception seen here, and then that null result is used to construct the Expression, which throws the error.

Fix is to also apply Invert for null comparison (which is a separate node, and not only comparison expressions), as well as only try to apply the optimization if the Invert returns a valid result (i.e. not null)